Added api and model generation from open-api spec Moved implementation code to 'impl' project
