+++ .travis.yml
@@ -3,7 +3,9 @@ addons:
   apt:
     packages:
       - oracle-java8-installer
-
+script:
+  - ./gradlew check
+  - ./gradlew codeCoverageReport
 before_cache:
   - rm -f  $HOME/.gradle/caches/modules-2/modules-2.lock
   - rm -fr $HOME/.gradle/caches/*/plugin-resolution/
