+++ build.gradle
@@ -18,12 +18,18 @@
 
 
 buildscript {
+    repositories {
+        mavenLocal()
+        mavenCentral()
+    }
+
     ext {
         kotlinVersion = '1.2.71'
         cordaVersion = '4.0'
         cordaGradlePluginsVersion = '4.0.25'
         springBootVersion = '2.1.3.RELEASE'
         springBootGradlePluginVersion = '2.1.3.RELEASE'
+        openApiPluginVersion = '3.3.4'
         nutsVersion = project.version
     }
 
@@ -33,12 +39,8 @@ buildscript {
         classpath "net.corda.plugins:cordapp:$cordaGradlePluginsVersion"
         classpath "net.corda.plugins:cordformation:$cordaGradlePluginsVersion"
         classpath "org.springframework.boot:spring-boot-gradle-plugin:$springBootGradlePluginVersion"
+        classpath "org.openapitools:openapi-generator-gradle-plugin:$openApiPluginVersion"
     }
-}
-
-plugins {
-    // Apply the Kotlin JVM plugin to add support for Kotlin on the JVM.
-    id 'org.jetbrains.kotlin.jvm' version '1.2.71'
 }
 
 ext {
@@ -48,35 +50,45 @@ ext {
     zeromqVersion = '0.5.0'
 }
 
-repositories {
-    mavenLocal()
-    mavenCentral()
-    jcenter()
-    maven {
-        url "https://ci-artifactory.corda.r3cev.com/artifactory/corda-releases/"
-    }
-}
-
-// exclude unneeded libs
-configurations {
-    implementation {
-        exclude group: 'com.github.corda.crash'
-        exclude group: 'com.github.bft-smart', module: 'library'
-        exclude group: 'ch.qos.logback'
-        exclude group: 'org.apache.logging.log4j', module: 'log4j-to-slf4j'
-    }
-
-    testImplementation {
-        exclude group: 'com.github.corda.crash'
-        exclude group: 'com.github.bft-smart', module: 'library'
-    }
-}
-
 allprojects {
-    apply plugin: 'jacoco'
+    repositories {
+        mavenLocal()
+        mavenCentral()
+        jcenter()
+        maven {
+            url "https://ci-artifactory.corda.r3cev.com/artifactory/corda-releases/"
+        }
+    }
+
+    apply plugin: 'org.jetbrains.kotlin.jvm'
+    apply plugin: 'kotlin'
     apply plugin: 'kotlin-spring'
-    apply plugin: 'org.springframework.boot'
     apply plugin: 'io.spring.dependency-management'
+    apply plugin: 'jacoco'
+
+    // exclude unneeded libs
+    configurations {
+        implementation {
+            exclude group: 'com.github.corda.crash'
+            exclude group: 'com.github.bft-smart', module: 'library'
+            exclude group: 'ch.qos.logback'
+            exclude group: 'org.apache.logging.log4j', module: 'log4j-to-slf4j'
+        }
+
+        testImplementation {
+            exclude group: 'com.github.corda.crash'
+            exclude group: 'com.github.bft-smart', module: 'library'
+        }
+    }
+
+    dependencies {
+        implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk8'
+
+        implementation("org.springframework.boot:spring-boot-starter-data-rest:${springBootVersion}")
+
+        testImplementation 'org.jetbrains.kotlin:kotlin-test'
+        testImplementation 'org.jetbrains.kotlin:kotlin-test-junit'
+    }
 
     tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
         kotlinOptions {
@@ -87,50 +99,62 @@ allprojects {
         }
     }
 
-    jacocoTestReport {
-        reports {
-            xml.enabled true
-            html.enabled true
-            csv.enabled false
-        }
+    test {
+        jvmArgs "-javaagent:${project.rootDir}/lib/quasar.jar",
+                "-Dco.paralleluniverse.fibers.verifyInstrumentation=true"
+        environment SPRING_PROFILES_ACTIVE: environment.SPRING_PROFILES_ACTIVE ?: "test"
     }
 }
 
-
-test {
-    jvmArgs "-javaagent:lib/quasar.jar",
-            "-Dco.paralleluniverse.fibers.verifyInstrumentation=true"
-}
-
-apply plugin: 'net.corda.plugins.cordapp'
-
 dependencies {
-    implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk8'
-
-    implementation("org.springframework.boot:spring-boot-starter:${springBootVersion}")
-    implementation "org.zeromq:jeromq:$zeromqVersion"
-    implementation("com.fasterxml.jackson.datatype:jackson-datatype-jsr310:$jacksonVersion")
-
-    cordaCompile("net.corda:corda-node-api:$cordaVersion")
-    cordaCompile("net.corda:corda-node:$cordaVersion")
-    cordaCompile("nl.nuts.consent:contract:$nutsVersion")
-    cordaCompile("nl.nuts.consent:flows:$nutsVersion")
-
-    implementation("org.apache.qpid:proton-j:0.27.3")
-
-    implementation "org.slf4j:jul-to-slf4j:$slf4jVersion"
-
-    testImplementation("net.corda:corda-test-utils:$cordaVersion")
-    testImplementation("net.corda:corda-node-driver:$cordaVersion")
-    testImplementation 'org.jetbrains.kotlin:kotlin-test'
-    testImplementation 'org.jetbrains.kotlin:kotlin-test-junit'
-
-    testImplementation group: 'org.apache.mina', name: 'mina-core', version: '2.1.0'
-    testImplementation group: 'javax.json', name: 'javax.json-api', version: '1.1.4'
-
-    testImplementation "org.springframework.boot:spring-boot-starter-test"
-
-    annotationProcessor "org.springframework.boot:spring-boot-configuration-processor"
+    implementation project(":impl")
 }
+
+task codeCoverageReport(type: JacocoReport) {
+    executionData fileTree(project.rootDir.absolutePath).include("**/build/jacoco/*.exec")
+
+    reports {
+        xml.enabled true
+        xml.destination file("${buildDir}/reports/jacoco/report.xml")
+        html.enabled true
+        html.destination file("${buildDir}/reports/jacoco/html")
+        csv.enabled false
+    }
+
+    subprojects.each {
+        sourceSets it.sourceSets.main
+    }
+}
+
+//change to dependsOn check when IT are added
+codeCoverageReport.dependsOn {
+    subprojects*.test
+}
+
+apply plugin: 'org.openapi.generator'
 
 check.dependsOn jacocoTestReport
+
+
+openApiGenerate {
+    generatorName = "kotlin-spring"
+    outputDir = "generated"
+    inputSpec = "https://nuts-node-rpc-spec.readthedocs.io/en/latest/_static/nuts-consent-bridge.yaml"
+    //inputSpec = "../nuts-node-rpc-spec/docs/_static/nuts-consent-bridge.yaml"
+    verbose = true
+    apiPackage = "nl.nuts.consent.bridge.api"
+    modelPackage = "nl.nuts.consent.bridge.model"
+    invokerPackage = "nl.nuts.consent.bridge"
+    configOptions = [
+            dateLibrary: "java8",
+            serviceInterface: true,
+            exceptionHandler: true,
+            gradleBuildFile: false,
+            supportingFiles: []
+    ]
+}
+
+openApiValidate {
+    //inputSpec = "../nuts-node-rpc-spec/docs/_static/nuts-consent-bridge.yaml"
+    inputSpec = "https://nuts-node-rpc-spec.readthedocs.io/en/latest/_static/nuts-consent-bridge.yaml"
+}
