+++ gradle.properties
@@ -1,4 +1,4 @@
 
-group=nl.nuts.consent
+group=nl.nuts.consent.bridge
 version=0.0.1
 cordapp.info.vendor = "Nuts community"
