+++ impl/src/main/resources/application.yaml
@@ -1,0 +1,10 @@
+spring:
+  application:
+    name: nutsConsentBridgeAPISpec
+
+  jackson:
+    serialization:
+      WRITE_DATES_AS_TIMESTAMPS: false
+
+server:
+  port: 8080
