+++ impl/src/test/kotlin/nl/nuts/consent/bridge/TestSerializationWhiteList.kt
