+++ impl/src/test/kotlin/nl/nuts/consent/bridge/api/ConsentApiIntegrationTest.kt
@@ -1,0 +1,67 @@
+/*
+ * Nuts consent bridge
+ * Copyright (C) 2019 Nuts community
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package nl.nuts.consent.bridge.api
+
+import com.nhaarman.mockito_kotlin.mock
+import com.nhaarman.mockito_kotlin.verify
+import nl.nuts.consent.bridge.model.EventStreamSetting
+import nl.nuts.consent.bridge.zmq.Publisher
+import nl.nuts.consent.bridge.zmq.Subscription
+import org.junit.Before
+import org.junit.Test
+import org.junit.runner.RunWith
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.boot.test.web.client.TestRestTemplate
+import org.springframework.http.HttpEntity
+import org.springframework.http.HttpHeaders
+import org.springframework.http.HttpStatus
+import org.springframework.http.MediaType
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner
+import org.springframework.test.util.ReflectionTestUtils
+import kotlin.test.assertEquals
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@RunWith(SpringJUnit4ClassRunner::class)
+class ConsentApiIntegrationTest {
+    @Autowired
+    lateinit var testRestTemplate: TestRestTemplate
+
+    @Autowired
+    lateinit var consentApiService: ConsentApiService
+
+    private var publisher: Publisher = mock()
+
+    @Before
+    fun setup() {
+        ReflectionTestUtils.setField(consentApiService, "publisher", publisher)
+    }
+
+    @Test
+    fun `POST to api consent event_stream adds subscription to publisher`() {
+        val headers = HttpHeaders()
+        headers.contentType = MediaType.APPLICATION_JSON
+        headers.accept = listOf(MediaType.TEXT_PLAIN)
+        val entity = HttpEntity(EventStreamSetting("topic", 1), headers)
+        val resp = testRestTemplate.postForEntity("/api/consent/event_stream", entity, String::class.java)
+        assertEquals(HttpStatus.OK, resp.statusCode) // part of protocol
+
+        verify(publisher).addSubscription(Subscription("topic", 1))
+    }
+}
