+++ impl/src/test/resources/META-INF/services/net.corda.core.serialization.SerializationWhitelist
