--- src/test/kotlin/nl/nuts/consent/bridge/rpc/test/DummyContract.kt
@@ -1,35 +1,0 @@
-/*
- * Nuts consent bridge
- * Copyright (C) 2019 Nuts community
- *
- *  This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-package nl.nuts.consent.bridge.rpc.test
-
-import net.corda.core.contracts.CommandData
-import net.corda.core.contracts.Contract
-import net.corda.core.transactions.LedgerTransaction
-
-class DummyContract : Contract {
-    companion object {
-        const val CONTRACT_ID = "nl.nuts.consent.bridge.rpc.test.DummyContract"
-    }
-
-    override fun verify(tx: LedgerTransaction) {
-        // OK
-    }
-
-    class DummyCommand : CommandData
-}
